functions/requirements.txt
azure-functions==1.0.8
azure-functions-worker==1.0.2
requests==2.22.0